Test delete error key

+++ src/Margo/Controller/AdminSubjectController.php
@@ -80,8 +80,15 @@ class AdminSubjectController
         if (!$subject) {
             $app->abort(404, 'The requested subject was not found.');
         }
-        $app['repository.subject']->delete($subject);
+        $o = 2;
+        if ( $o == 2){
+            $message = 'Ce cours est déjà attribué à un prof';
+            $app['session']->getFlashBag()->add('error', $message);
+        } else {
+            $app['repository.subject']->delete($subject);
+        }
         return $app->redirect($app['url_generator']->generate('admin_subjects'));
+
     }
 
 }
